[portfolio] Extract image utilities and animations
