+++ portfolio/components/ui/Figures/PhotoReceiptBoundingBox.tsx
@@ -8,6 +8,11 @@ import {
   type Point as ApiPoint,
 } from "../../../types/api";
 import { useSpring, useTransition, animated } from "@react-spring/web";
+import AnimatedLineBox from "../animations/AnimatedLineBox";
+import {
+  detectImageFormatSupport,
+  getBestImageUrl,
+} from "../../../utils/imageFormat";
 import {
   computeHullCentroid,
   computeReceiptBoxFromLineEdges,
@@ -15,6 +20,7 @@ import {
 } from "../../../utils/geometry";
 
 // Define simple point and line-segment shapes
+const isDevelopment = process.env.NODE_ENV === "development";
 type Point = ApiPoint;
 type LineSegment = {
   x1: number;
@@ -24,119 +30,7 @@ type LineSegment = {
   key: string;
 };
 
-const isDevelopment = process.env.NODE_ENV === "development";
 
-// Browser format detection utilities
-const detectImageFormatSupport = (): Promise<{
-  supportsAVIF: boolean;
-  supportsWebP: boolean;
-}> => {
-  return new Promise((resolve) => {
-    const userAgent = navigator.userAgent;
-
-    // Safari version detection
-    const getSafariVersion = (): number | null => {
-      if (userAgent.includes("Chrome")) return null; // Chrome has Safari in UA, exclude it
-
-      const safariMatch = userAgent.match(/Version\/([0-9.]+).*Safari/);
-      if (safariMatch) {
-        return parseFloat(safariMatch[1]);
-      }
-      return null;
-    };
-
-    const isChrome =
-      userAgent.includes("Chrome") && userAgent.includes("Google Chrome");
-    const isFirefox = userAgent.includes("Firefox");
-    const safariVersion = getSafariVersion();
-    const isSafari = safariVersion !== null;
-
-    // WebP support detection
-    let supportsWebP = false;
-
-    if (isChrome || isFirefox) {
-      // Chrome and Firefox have excellent WebP support
-      supportsWebP = true;
-    } else if (isSafari && safariVersion && safariVersion >= 14) {
-      // Safari 14+ supports WebP (macOS Big Sur, iOS 14)
-      supportsWebP = true;
-    } else {
-      // Try canvas test as fallback
-      try {
-        const canvas = document.createElement("canvas");
-        canvas.width = 1;
-        canvas.height = 1;
-        const ctx = canvas.getContext("2d");
-        if (ctx) {
-          const webpDataUrl = canvas.toDataURL("image/webp", 0.5);
-          supportsWebP = webpDataUrl.indexOf("data:image/webp") === 0;
-        }
-      } catch (error) {
-        supportsWebP = false;
-      }
-    }
-
-    // AVIF support detection
-    const detectAVIF = (): Promise<boolean> => {
-      if (isChrome) {
-        // Chrome 85+ supports AVIF (September 2020)
-        const chromeMatch = userAgent.match(/Chrome\/([0-9]+)/);
-        if (chromeMatch && parseInt(chromeMatch[1]) >= 85) {
-          return Promise.resolve(true);
-        }
-      }
-
-      if (isFirefox) {
-        // Firefox 93+ supports AVIF (October 2021)
-        const firefoxMatch = userAgent.match(/Firefox\/([0-9]+)/);
-        if (firefoxMatch && parseInt(firefoxMatch[1]) >= 93) {
-          return Promise.resolve(true);
-        }
-      }
-
-      if (isSafari && safariVersion && safariVersion >= 16.4) {
-        // Safari 16.4+ supports AVIF (March 2023)
-        return Promise.resolve(true);
-      }
-
-      // Fallback to image test for unknown browsers or older versions
-      return new Promise<boolean>((resolveAVIF) => {
-        const img = new Image();
-        img.onload = () => resolveAVIF(true);
-        img.onerror = () => resolveAVIF(false);
-        img.src =
-          "data:image/avif;base64,AAAAIGZ0eXBhdmlmAAAAAGF2aWZtaWYxbWlhZk1BMUIAAADybWV0YQAAAAAAAAAoaGRscgAAAAAAAAAAcGljdAAAAAAAAAAAAAAAAGxpYmF2aWYAAAAADnBpdG0AAAAAAAEAAAAeaWxvYwAAAABEAAABAAEAAAABAAABGgAAAB0AAAAoaWluZgAAAAAAAQAAABppbmZlAgAAAAABAABhdjAxQ29sb3IAAAAAamlwcnAAAABLaXBjbwAAABRpc3BlAAAAAAAAAAEAAAABAAAAEHBpeGkAAAAAAwgICAAAAAxhdjFDgQ0MAAAAABNjb2xybmNseAACAAIAAYAAAAAXaXBtYQAAAAAAAAABAAEEAQKDBAAAACVtZGF0EgAKCBgABogQEAwgMg8f8D///8WfhwB8+ErK42A=";
-      });
-    };
-
-    detectAVIF().then((supportsAVIF) => {
-      resolve({ supportsAVIF, supportsWebP });
-    });
-  });
-};
-
-// Get the best available image URL based on browser support and available formats
-const getBestImageUrl = (
-  image: ImageType,
-  formatSupport: { supportsAVIF: boolean; supportsWebP: boolean }
-): string => {
-  const baseUrl = isDevelopment
-    ? "https://dev.tylernorlund.com"
-    : "https://www.tylernorlund.com";
-
-  // Try AVIF first (best compression)
-  if (formatSupport.supportsAVIF && image.cdn_avif_s3_key) {
-    return `${baseUrl}/${image.cdn_avif_s3_key}`;
-  }
-
-  // Try WebP second (good compression, wide support)
-  if (formatSupport.supportsWebP && image.cdn_webp_s3_key) {
-    return `${baseUrl}/${image.cdn_webp_s3_key}`;
-  }
-
-  // Fallback to JPEG (universal support)
-  return `${baseUrl}/${image.cdn_s3_key}`;
-};
 
 /**
  * Compute the convex hull of a set of points using Graham scan algorithm
@@ -521,78 +415,6 @@ const estimateReceiptPolygonFromLines = (lines: Line[]) => {
   };
 };
 
-// AnimatedLineBox: already defined for words
-interface AnimatedLineBoxProps {
-  line: Line; // Adjust type as needed
-  svgWidth: number;
-  svgHeight: number;
-  delay: number;
-}
-
-const AnimatedLineBox: React.FC<AnimatedLineBoxProps> = ({
-  line,
-  svgWidth,
-  svgHeight,
-  delay,
-}) => {
-  // Convert normalized coordinates to absolute pixel values.
-  const x1 = line.top_left.x * svgWidth;
-  const y1 = (1 - line.top_left.y) * svgHeight;
-  const x2 = line.top_right.x * svgWidth;
-  const y2 = (1 - line.top_right.y) * svgHeight;
-  const x3 = line.bottom_right.x * svgWidth;
-  const y3 = (1 - line.bottom_right.y) * svgHeight;
-  const x4 = line.bottom_left.x * svgWidth;
-  const y4 = (1 - line.bottom_left.y) * svgHeight;
-  const points = `${x1},${y1} ${x2},${y2} ${x3},${y3} ${x4},${y4}`;
-
-  // Compute the polygon's centroid.
-  const centroidX = (x1 + x2 + x3 + x4) / 4;
-  const centroidY = (y1 + y2 + y3 + y4) / 4;
-
-  // Animate the polygon scaling from 0 to 1, with the centroid as the origin.
-  const polygonSpring = useSpring({
-    from: { transform: "scale(0)" },
-    to: { transform: "scale(1)" },
-    delay: delay,
-    config: { duration: 800 },
-  });
-
-  // Animate the centroid marker:
-  // 1. Fade in at the computed centroid.
-  // 2. Then animate its y coordinate to mid‑Y.
-  const centroidSpring = useSpring({
-    from: { opacity: 0, cy: centroidY },
-    to: async (next) => {
-      await next({ opacity: 1, cy: centroidY, config: { duration: 300 } });
-      await next({ cy: svgHeight / 2, config: { duration: 800 } });
-    },
-    delay: delay + 30,
-  });
-
-  return (
-    <>
-      <animated.polygon
-        style={{
-          ...polygonSpring,
-          transformOrigin: "50% 50%",
-          transformBox: "fill-box",
-        }}
-        points={points}
-        fill="none"
-        stroke="var(--color-red)"
-        strokeWidth="2"
-      />
-      <animated.circle
-        cx={centroidX}
-        cy={centroidSpring.cy}
-        r={10}
-        fill="var(--color-red)"
-        style={{ opacity: centroidSpring.opacity }}
-      />
-    </>
-  );
-};
 
 // AnimatedConvexHull: component for animating convex hull calculation
 interface AnimatedConvexHullProps {
